feat: :recycle: Reordenar endpoints en la API y solucionar problemas, Reordenar endpoints en la API y solucionar problemas
Se realizo la consolidacion de las publicaciones y reordenamiento de controladores junto con los endpoints y la refactorizacion del cliente Supabase en todos los microservicios
