+++ apps/auth/src/routes/auth.ts
@@ -1,5 +1,5 @@
 import { Router } from "express"
-import { login, register } from "../controllers/auth"
+import { login, register } from "../controllers/request"
 
 const router = Router()
