+++ apps/entities/src/index.ts
@@ -2,7 +2,6 @@ import express from "express"
 import cors from "cors"
 import helmet from "helmet"
 import morgan from "morgan"
-import { errorHandler, ENTITIES_PORT, getHeaders } from "@repo/utils"
 import "dotenv/config"
 import giverRequestRouter from "./routes/giverRequest"
