+++ apps/entities/src/routes/giverRequest.ts
@@ -1,9 +1,9 @@
 import { Router } from "express"
 import { requireRole } from "@repo/utils"
-import { listGiverRequests } from "../controllers/giverRequests"
+import { listGiverRequests } from "../controllers/request"
 
 const router = Router()
 
-router.get("/giverRequests", requireRole(19), listGiverRequests)
+router.get("/giver-Requests", requireRole(19), listGiverRequests)
 
 export default router
