+++ apps/gateway/src/docs/openapi.ts
@@ -1,10 +1,19 @@
 import { OpenAPIRegistry, OpenApiGeneratorV3 } from "@asteasolutions/zod-to-openapi"
 import { registerAuthPaths } from "./endpoints/auth"
+import { mineRequestDocs, ConfirmAcceptDocs, validateCodeDocs } from "./endpoints/adoptions"
+import { giverRequestDocs, adoptionHistory } from "./endpoints/entities"
 
 export function buildOpenApi() {
     const registry = new OpenAPIRegistry()
 
     registerAuthPaths(registry)
+
+    mineRequestDocs(registry)
+    ConfirmAcceptDocs(registry)
+    validateCodeDocs(registry)
+
+    giverRequestDocs(registry)
+    adoptionHistory(registry)
 
     const generator = new OpenApiGeneratorV3(registry.definitions)
