--- apps/media/src/controllers/giverRequest.ts
@@ -1,57 +1,0 @@
-import { Request, Response, NextFunction } from "express"
-import path from "path"
-import fs from "fs/promises"
-
-import { getAllFiles, UPLOADS_BASE } from "../utils"
-import { HttpError } from "../middleware/upload"
-import { AppResponse } from "@repo/utils"
-
-export const giverPost = (req: Request, res: Response, next: NextFunction) => {
-    try {
-        const { rut } = req.params
-
-        const many = (req.files as Express.Multer.File[]) ?? []
-        const one = req.file as Express.Multer.File | undefined
-        const files = many.length ? many : one ? [one] : []
-
-        if (!files.length) throw new HttpError(400, "No files were provided")
-
-        const uploaded = files.map((file) => ({
-            url: `/uploads/account-request/${rut}/${file.filename}`,
-            fileName: file.filename,
-            size: file.size,
-            mime: file.mimetype,
-        }))
-
-        return AppResponse(res, 201, "Files uploaded successfully", uploaded)
-    } catch (err) {
-        next(err)
-    }
-}
-
-export const getGiverFiles = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-        const filesPath = path.join(UPLOADS_BASE, "account-request")
-        await fs.access(filesPath).catch(() => {
-            throw new HttpError(404, "No files found for this entity type")
-        })
-
-        const allFiles = await getAllFiles(filesPath)
-        const grouped: Record<string, string[]> = {}
-
-        for (const abs of allFiles) {
-            const rel = path.relative(path.join(__dirname, ".."), abs)
-            const url = `/${rel.replace(/\\/g, "/")}`
-
-            const parts = url.split("/")
-            const entityId = parts[3]
-
-            if (!grouped[entityId]) grouped[entityId] = []
-            grouped[entityId].push(url)
-        }
-
-        return AppResponse(res, 200, "Files retrieved successfully", grouped)
-    } catch (err) {
-        next(err)
-    }
-}
